// client/src/core/store/actionTypes.ts
export enum Actions {
   RIGHTSIDEBAR_CHANGE_VIEW = "RIGHTSIDEBAR_CHANGE_VIEW",
   LEFTSIDEBAR_CHANGE_VIEW = "LEFTSIDEBAR_CHANGE_VIEW",
   ACCOUNT_SETTINGS_UPDATE = "ACCOUNT_SETTINGS_UPDATE",
   ACCOUNT_PASSWORD_CHANGE = "ACCOUNT_PASSWORD_CHANGE",
   AUTH_SIGN_IN = "AUTH_SIGN_IN",
   AUTH_LOGOUT = "AUTH_LOGOUT",
   CHATS_SELECT_CHAT = "CHATS_SELECT_CHAT",
   CHATS_UPDATE_CHAT = "CHATS_UPDATE_CHAT",
   CHATS_UPLOAD_CHATS = "CHATS_UPLOAD_CHATS",
   SELECTED_CHAT_CHANGE = "SELECTED_CHAT_CHANGE",
   TOKENS_SAVE = "TOKENS_SAVE",
   MESSENGER_SAVE = "MESSENGER_SAVE",
   USERS_SAVE = "USERS_SAVE",
}